investment logic, needs fixing

+++ crud-bank-app/src/main/java/com/ironhack/crudbankapp/model/CheckingAccount.java
@@ -4,6 +4,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Inheritance;
 import jakarta.persistence.InheritanceType;
 
+import java.math.BigDecimal;
+
 @Entity
 @Inheritance(strategy= InheritanceType.JOINED)
 public class CheckingAccount extends Account {
@@ -16,12 +18,12 @@ public class CheckingAccount extends Account {
         super(owner);
     }
     @Override
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return super.getBalance();
     }
 
     @Override
-    public void setBalance(Double balance) {
+    public void setBalance(BigDecimal balance) {
         super.setBalance(balance);
     }
